feat: :sparkles: added attribute widget for admin
