+++ src/widgets/attributeWidget/index.ts
